+++ BE/src/main/java/ifive/idrop/driver/service/DriverService.java
@@ -13,8 +13,9 @@ import ifive.idrop.fcm.AlarmMessage;
 import ifive.idrop.fcm.NotificationUtill;
 import ifive.idrop.driver.repository.DriverRepository;
 import ifive.idrop.parent.domain.Parent;
+import ifive.idrop.pickup.domain.PickUp;
 import ifive.idrop.repository.NotificationRepository;
-import ifive.idrop.repository.PickUpRepository;
+import ifive.idrop.pickup.repository.PickUpRepository;
 import ifive.idrop.util.RequestSchedule;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
