+++ BE/src/main/java/ifive/idrop/entity/PickUpInfo.java
@@ -2,6 +2,7 @@ package ifive.idrop.entity;
 
 import ifive.idrop.driver.domain.Driver;
 import ifive.idrop.parent.domain.Parent;
+import ifive.idrop.pickup.domain.PickUp;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.AllArgsConstructor;
