+++ BE/src/main/java/ifive/idrop/parent/service/ParentService.java
@@ -17,7 +17,8 @@ import ifive.idrop.fcm.NotificationUtill;
 import ifive.idrop.driver.repository.DriverRepository;
 import ifive.idrop.parent.domain.Parent;
 import ifive.idrop.parent.repository.ParentRepository;
-import ifive.idrop.repository.PickUpRepository;
+import ifive.idrop.pickup.domain.PickUp;
+import ifive.idrop.pickup.repository.PickUpRepository;
 import lombok.RequiredArgsConstructor;
 import org.json.JSONException;
 import org.springframework.stereotype.Service;
