+++ BE/src/main/java/ifive/idrop/pickup/domain/PickUp.java
@@ -1,6 +1,8 @@
-package ifive.idrop.entity;
+package ifive.idrop.pickup.domain;
 
 import ifive.idrop.driver.domain.Driver;
+import ifive.idrop.entity.Child;
+import ifive.idrop.entity.PickUpInfo;
 import ifive.idrop.parent.domain.Parent;
 import jakarta.persistence.*;
 import lombok.*;
